+++ IronNet/src/ironapps/vuln_server.h
@@ -1,0 +1,6 @@
+#ifndef IRON_VULN_SERVER_H
+#define IRON_VULN_SERVER_H
+
+int vuln_server_start(void);
+
+#endif